Use scales instead of production rules for stroke and size
This is more in line with the Vega way of thinking and does make it
pretty clear how different sites are styled since the maps are explicit
in the spec.

It's also a requirement for drawing useful legends in Vega, although
currently I'm not adding a legend because it's impossible to position it
on the bottom of the viz.  I don't want to eat up horizontal space since
that's the primary data axis.