Tally partially converted CpH sites indicated as YpH
As requested by Dylan.  Visually these sites are represented as
half-filled red circles, paralleling the behaviour for YpG sites.

The visualization gains a new scale for stroke width, which makes for
nicer looking red circles.  The explanatory text about ambiguous sites
is reorganized a bit since it has grown with this change.  Now that more
than just cytosines are reported as failed conversion sites, each site
is prefixed with its nucleic acid code in the summary spreadsheet.  This
was specifically asked for by Dylan.