Refer to the methylation "viz" as a "diagram" instead
"Viz", for "visualization", was a short and handy moniker, but never sat
quite right with me since it was so general.  "Diagram" is also general,
but encompasses less kinds of things and brings to mind line drawings in
particular (which is what we have).  Unlike "viz", "diagram" doesn't
make me think of plots, charts, heatmaps, or other visualizations of
data.  Since I'm working on adding another visualization (a heatmap), it
seemed time to fix the way the code refers to the existing one.