Rework the diagram component to re-use the same Vega view
Instead of re-parsing the spec and re-creating the view every time the
data changed.  This is more parsimonious, but the primary upshot is that
the view isn't temporarily removed from the DOM before being replaced,
and thus doesn't cause a flash of content or force the viewport to
scroll to the top.  These side-effects weren't nearly as noticeable
before the page required scrolling.

Another upshot is that a bunch of other callbacks can directly close
over the view object and know that it must exist rather than looking for
it on the element and checking if it (and the element itself!) still
exist.