Optionally label CpG sites with their ordinal number or nothing at all
Instead of only labeling with alignment position.

This also includes the ability to add a fixed offset to the labels so
that the coordinate space of alignment positions can be shifted to match
HXB2.  (This only works if your alignment doesn't have any indels
relative to HXB2, but Dylan's don't.)

Requested by Dylan and Jim.