Patch Vega 2.6.5 to fix a data streaming bug
I ran into two bugs while trying to support grouping sequences visually
by a dynamic field.  In the process of debugging, I patched one bug with
an obvious fix.  While it may not be the most ideal fix, it allowed me
to move forward with the help of a workaround for the second bug.  The
resulting feature will come in subsequent commits.

The patch guards two places where, for some reason, the data item Vega
wants to remove from a data flow propagation is already gone.  Without
the guard, the code blows up because of property accesses on `item` when
`item == null`.

This change switches us from a minified Vega 2.6.5 to using a patched,
unminified Vega 2.6.5.  The patch is included alongside the JS for
posterity in case we need to re-apply it.

Some day we'll port this to Vega 3 and hopefully leave the bugs behind.
(Or just use plain D3 scaling with native SVG generation in Angular.)