Summarize groups of sequences (or all sequences) in the diagram
The diagram can now function similarly to the heatmap, although the
visual representation is very different.  This approach was worked out
with Dylan's and Jim's guidance as they prepared figures for a progress
report.

The original prototypes there were done with a grayscale color ramp for
methylation level, but that was deemed to be unsuitably hard to discern.
In this polishing of the prototype, I've opted for a quantized scale of
line-shaded dots.

The driving idea behind the implementation is to feed the diagram
synthetic sequence data based on the heatmap aggregation.