wip! color ramp instead of a graphic ramp
This is the color ramp Dylan used for his HIV Dynamics poster.  Ideally,
I think, the ramp to use becomes an option instead of entirely replacing
the graphic ramp with a color ramp.

Note that this doesn't bother changing the hover highlight ramp, which
should also be updated in any final version.